requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
python-decouple==3.8
python-dotenv==1.2.2
six==1.17.0
